feat(store): add provideStore and provideState functions for standalone APIs
